requirements.txt
asgiref==3.9.2
attrs==25.4.0
cachetools==5.5.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
cryptography==41.0.7
defusedxml==0.7.1
Django==4.2
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.10.1
google-auth==2.35.0
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
oauthlib==3.3.1
packaging==25.0
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
PyJWT==2.8.0
python-decouple==3.8
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
requests==2.31.0
requests-oauthlib==2.0.0
rpds-py==0.28.0
rsa==4.9.1
setuptools==80.9.0
sqlparse==0.5.3
uritemplate==4.2.0
urllib3==2.5.0
wheel==0.45.1
whitenoise==6.11.0
